version a produccion

+++ castellano/PHP/EnDirElimina.php
@@ -1,11 +1,17 @@
 <?php
 include($_SERVER['DOCUMENT_ROOT']."/rao/rao_con.php");
 
-$id = $_GET["id"];
+session_start();
 
-$SQL = "DELETE FROM EnDirHome WHERE IdEnDirHome =".$id;
-if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
+if ($_SESSION["Creacio"]){
 
 
-echo $id;
+	$id = mysqli_real_escape_string($mysqli,$_GET["id"]);
+
+	$SQL = "DELETE FROM EnDirHome WHERE IdEnDirHome =".$id;
+	if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
+
+
+	echo $id;
+}
 ?>
