+++ castellano/PHP/LinMDDelete.php
@@ -1,20 +1,26 @@
 <?php
 include($_SERVER['DOCUMENT_ROOT']."/rao/rao_con.php");
 
-$IdLinMD = $_GET["IdLinMD"];
+session_start();
 
-$SQL = "SELECT * FROM LinMD WHERE IdLinMD = " . $IdLinMD;
-if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
+if ($_SESSION["Creacio"]){
 
 
- while ($row = $result->fetch_assoc())
-{
-	$IdCap = $row["IdCapMenu"];
+	$IdLinMD = mysqli_real_escape_string($mysqli,$_GET["IdLinMD"]);
+
+	$SQL = "SELECT * FROM LinMD WHERE IdLinMD = " . $IdLinMD;
+	if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
+
+
+	 while ($row = $result->fetch_assoc())
+	{
+		$IdCap = $row["IdCapMenu"];
+	}
+
+	$SQL = "DELETE FROM LinMD WHERE IdLinMD = " . $IdLinMD; 
+	if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
+
+
+	echo $IdCap;//."|".$IdLin;
 }
-
-$SQL = "DELETE FROM LinMD WHERE IdLinMD = " . $IdLinMD; 
-if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
-
-
-echo $IdCap;//."|".$IdLin;
 ?>
