+++ castellano/PHP/LinPageDelete.php
@@ -1,12 +1,18 @@
 <?php
 include($_SERVER['DOCUMENT_ROOT']."/rao/rao_con.php");
 
-$idCap = $_GET["IdCap"];
-$IdLin = $_GET["IdLin"];
+session_start();
 
-$SQL = "DELETE FROM LinMenu WHERE IdLinMenu = " . $IdLin; 
-if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
+if ($_SESSION["Creacio"]){
 
 
-echo $idCap;//."|".$IdLin;
+	$idCap = mysqli_real_escape_string($mysqli,$_GET["IdCap"]);
+	$IdLin = mysqli_real_escape_string($mysqli,$_GET["IdLin"]);
+
+	$SQL = "DELETE FROM LinMenu WHERE IdLinMenu = " . $IdLin; 
+	if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
+
+
+	echo $idCap;//."|".$IdLin;
+}
 ?>
