+++ castellano/PHP/UserGUUDelete.php
@@ -1,12 +1,17 @@
 <?php
 include($_SERVER['DOCUMENT_ROOT']."/rao/rao_con.php");
 
-$id = $_GET["id"];
+session_start();
 
-$SQL = "DELETE FROM Users WHERE IdUser = $id";
-if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
+if ($_SESSION["Creacio"]){
 
 
+	$id = mysqli_real_escape_string($mysqli,$_GET["id"]);
+
+	$SQL = "DELETE FROM Users WHERE IdUser = $id";
+	if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
+
+}
 
 //echo $SQL;
 ?>
