+++ english/PHP/CategoriaDirectoriDelete.php
@@ -1,10 +1,17 @@
 <?php
 include($_SERVER['DOCUMENT_ROOT']."/rao/rao_con.php");
 
-$IdCat = $_GET["IdCat"];
+session_start();
 
-$SQL = "DELETE FROM DirectoriCategoria WHERE IdDirectoriCategoria = " . $IdCat; 
-if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli)); 
+if ($_SESSION["Creacio"]){
 
-echo $idCap;//."|".$IdLin;
+
+	$IdCat = mysqli_real_escape_string($mysqli,$_GET["IdCat"]);
+
+	$SQL = "DELETE FROM DirectoriCategoria WHERE IdDirectoriCategoria = " . $IdCat; 
+	if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli)); 
+
+	echo $idCap;//."|".$IdLin;
+}
+
 ?>
