+++ english/PHP/DestacatsElimina.php
@@ -1,17 +1,23 @@
 <?php
 include($_SERVER['DOCUMENT_ROOT']."/rao/rao_con.php");
 
-$id = $_GET["id"];
-$IMG = $_GET["IMG"];
+session_start();
 
-if ($IMG)
-{
-	unlink('../ImgDes/'.$IMG);	
+if ($_SESSION["Creacio"]){
+
+
+	$id = mysqli_real_escape_string($mysqli,$_GET["id"]);
+	$IMG = mysqli_real_escape_string($mysqli,$_GET["IMG"]);
+
+	if ($IMG)
+	{
+		unlink('../ImgDes/'.$IMG);	
+	}
+
+	$SQL = "DELETE FROM Destacat WHERE IdDestacat =".$id;
+	if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
+
+
+	echo $id;
 }
-
-$SQL = "DELETE FROM Destacat WHERE IdDestacat =".$id;
-if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
-
-
-echo $id;
 ?>
