+++ english/PHP/Directori/DirectoriDelete.php
@@ -2,22 +2,27 @@
 include($_SERVER['DOCUMENT_ROOT']."/rao/rao_con.php");
 include($_SERVER['DOCUMENT_ROOT']."/rao/PonQuita.php");
 
-$id = Pon($_POST["id"]);
+session_start();
 
-$SQL = "SELECT Imatge FROM Directori WHERE IdDirectori = ".$id;	
-if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
+if ($_SESSION["Creacio"]){
 
 
- while ($row = $result->fetch_assoc())
-{
-	if(file_exists("../../../imgDirectori/".$row["Imatge"])) unlink("../../../imgDirectori/".$row["Imatge"]);
+	$id = Pon(mysqli_real_escape_string($mysqli,$_POST["id"]));
+
+	$SQL = "SELECT Imatge FROM Directori WHERE IdDirectori = ".$id;	
+	if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
+
+
+	 while ($row = $result->fetch_assoc())
+	{
+		if(file_exists("../../../imgDirectori/".$row["Imatge"])) unlink("../../../imgDirectori/".$row["Imatge"]);
+	}
+
+
+	$SQL = "DELETE FROM Directori WHERE IdDirectori = ".$id;
+
+	if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
+
+
 }
-
-
-$SQL = "DELETE FROM Directori WHERE IdDirectori = ".$id;
-
-if (!$result = $mysqli->query($SQL))printf("Errormessage: %s\n", mysqli_error($mysqli));
-
-
-
 ?>
